+++ NEWS.md
@@ -1,4 +1,4 @@
-# medicalcoder 0.8.0.9000
+# medicalcoder 0.8.1
 
 ## Bug Fixes